Fix upload file account and CI/CD

+++ modules/account/http/handler.go
@@ -27,11 +27,15 @@ func NewAccountHandler(service service.AccountService) *AccountHandler {
 func (h *AccountHandler) Import(c *fiber.Ctx) error {
 	// Validate the uploaded file
 	if err := validation.ValidateAccountRequest(c); err != nil {
-		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": err.Error()})
+		return err // Error response is already handled in the validation function
 	}
 
-	// Save file temporarily
-	file, _ := c.FormFile("file_csv")
+	// Get the file from the request
+	file, err := c.FormFile("file_csv")
+	if err != nil {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "File is required"})
+	}
+
 	tempPath := "./temp/" + file.Filename
 
 	// Ensure the temp directory exists
@@ -47,6 +51,9 @@ func (h *AccountHandler) Import(c *fiber.Ctx) error {
 
 	// Retrieve user_id from the form
 	userID := c.FormValue("user_id")
+	if userID == "" {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "User ID is required"})
+	}
 
 	go func() {
 		defer os.Remove(tempPath)
@@ -108,5 +115,8 @@ func (h *AccountHandler) Import(c *fiber.Ctx) error {
 		}
 	}()
 
-	return c.JSON(fiber.Map{"message": "File processed successfully"})
+	return c.JSON(fiber.Map{
+		"message": "File upload successful, processing in background",
+		"status":  "success",
+	})
 }
